+++ src/CoordinationId.php
@@ -6,13 +6,8 @@ namespace byrokrat\id;
 
 use byrokrat\id\Helper\DateTimeCreator;
 use byrokrat\id\Helper\NumberParser;
+use byrokrat\id\Exception\UnableToCreateIdException;
 
-/**
- * Coordination id number
- *
- * A coordination number is like a personal number except that 60 is added
- * to the date of birth.
- */
 class CoordinationId extends PersonalId
 {
     /**
@@ -20,6 +15,20 @@ class CoordinationId extends PersonalId
      */
     private $datestr;
 
+    /**
+     * Create coordination id number
+     *
+     * A coordination number is like a personal number except that 60 is added
+     * to the date of birth.
+     *
+     * NOTE that coordination numbers may contain a date that does not actually
+     * exist. The month and day parts may on some cases be set to '00'. And
+     * the day part may in some cases refer to undefined dates, such as the 30th
+     * of february. In these cases the parser will convert the date to a valid
+     * date, even though it may not reflect the actual birth date.
+     *
+     * @throws UnableToCreateIdException On failure to create id
+     */
     public function __construct(string $raw)
     {
         list($century, $this->datestr, $delim, $serialPost, $check) = NumberParser::parse(self::PATTERN, $raw);