+++ src/Exception/InvalidCheckDigitException.php
@@ -5,6 +5,6 @@ namespace byrokrat\id\Exception;
 /**
  * Exception thrown when trying to create on id with invalid check digit
  */
-class InvalidCheckDigitException extends RuntimeException
+class InvalidCheckDigitException extends UnableToCreateIdException
 {
 }