+++ src/Exception/InvalidStructureException.php
@@ -5,6 +5,6 @@ namespace byrokrat\id\Exception;
 /**
  * Exception thrown when supplied id is not valid
  */
-class InvalidStructureException extends RuntimeException
+class InvalidStructureException extends UnableToCreateIdException
 {
 }